==> WorkshopCodes/Shaastra2026-IITMadras/04-NLP/07-NamedEntityRecognition.ipynb <==
import pandas as pd
import spacy
from spacy import displacy
import matplotlib.pyplot as plt
from pyvis.network import Network

# pip install networkx pyvis

# 01 - Load Data

# data source: Kaggle - https://www.kaggle.com/datasets/naseralqaydeh/named-entity-recognition-ner-corpus
df = pd.read_csv("../Data/ner.csv")
df.head()

# Load spaCy's pre-trained English model
nlp = spacy.load("en_core_web_sm")

# Function to extract named entities from a text
def extract_entities(text):
    doc = nlp(text)
    entities = [(ent.text, ent.label_) for ent in doc.ents]
    return entities

df100 = df.head(100)

# Apply NER on the text column
df100['entities'] = df100['Sentence'].apply(extract_entities)

# Save the results to a new CSV
df100.to_csv("data_with_entities.csv", index=False)

print("NER extraction completed. Results saved to 'data_with_entities.csv'.")


df100.head()

# Analysis

### Look at most common types of entities

ent_dict = {}
label_dict = {}
label_ent_dict = {}
for i in range(len(df100)):
    entlist = df100['entities'].iloc[i]
    for ent in entlist:
        if ent in ent_dict.keys():
            ent_dict[ent] += 1
        else:
            ent_dict[ent] = 1
        nent, label = ent
        if label in label_dict.keys():
            label_dict[label] += 1
        else:
            label_dict[label] = 1
        if label in label_ent_dict.keys():
            label_ent_dict[label].append(nent)
        else:
            label_ent_dict[label] = [nent]

for k in label_ent_dict.keys():
    label_ent_dict[k] = set(label_ent_dict[k])

# One sample entry
df100['entities'].iloc[99]

# Count word-label frequencies
{w:ent_dict[w] for w in sorted(ent_dict, key=ent_dict.get, reverse=True)}

# Count label frequencies
{w:label_dict[w] for w in sorted(label_dict, key=label_dict.get, reverse=True)}

label_ent_dict

# Display Entities

# Display entities
text = df100['Sentence'].iloc[0]
doc = nlp(text)
displacy.render(doc, style="ent")

# Get relations between labels
label_list = []
for i in range(len(df100)):
    entlist = df100['entities'].iloc[i]
    label_ll = []
    for ent in entlist:
        nent, label = ent
        label_ll.append(label)
    label_list.append(label_ll)

label_list

from itertools import combinations 

def item_pairs(inp_list):
    return list(combinations(sorted(inp_list), 2))    

A = ['cat', 'baby', 'apple']
item_pairs(A)

label_pairs= {}
for ll in label_list:
    pairs = item_pairs(ll)
    for pp in pairs:
        if pp in label_pairs.keys():
            label_pairs[pp] += 1
        else:
            label_pairs[pp] = 1

{w:label_pairs[w] for w in sorted(label_pairs, key = label_pairs.get, reverse=True)}

# Get relations between entities
entity_list = []
for i in range(len(df100)):
    entlist = df100['entities'].iloc[i]
    ent_ll = []
    for ent in entlist:
        nent, label = ent
        if label not in ['CARDINAL', 'DATE']:
            ent_ll.append(nent)
    entity_list.append(ent_ll)

entity_list

# Create a dictionary 
entity_pairs= {}
for ents in entity_list:
    pairs = item_pairs(ents)
    for pp in pairs:
        if pp in entity_pairs.keys():
            entity_pairs[pp] += 1
        else:
            entity_pairs[pp] = 1

{w:entity_pairs[w] for w in sorted(entity_pairs, key = entity_pairs.get, reverse=True)}

#net = Network()
net = Network(height="750px", width="100%", bgcolor="#ffffff", font_color="black")

for key, vals in entity_pairs.items():
    a, b = key
    net.add_node(a)
    net.add_node(b)
    net.add_edge(a, b, weight=vals)

net.show('graph.html', notebook=False)

#net = Network()
net = Network(height="750px", width="100%", bgcolor="#ffffff", font_color="black")

net.add_node('Cap')
net.add_node('Cap')
net.add_node('Tap')
net.add_edge('Cap', 'Tap', weight=vals)

net.show('graphz.html', notebook=False)

# Draw insights about specific entities

# Q1: Do the news articles report positively about London? 

# TextBlob
from textblob import TextBlob

df100['sentiment'] = df100['Sentence'].apply(lambda x: TextBlob(x).sentiment.polarity)

df100.head()

dff = df100[df100['Sentence'].str.contains('Afghanistan')]

dff

dff['sentiment'].describe()

dff = df100[df100['Sentence'].str.contains('Mogadishu')]
dff['sentiment'].describe()

dff = df100[df100['Sentence'].str.contains('London')]
dff['sentiment'].describe()
