==> WorkshopCodes/Shaastra2026-IITMadras/04-NLP/08-NER-RelationshipExtraction.ipynb <==
import pandas as pd
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM
import torch
import re

# pip install networkx pyvis

# 01 - Load Data

# data source: Kaggle - https://www.kaggle.com/datasets/naseralqaydeh/named-entity-recognition-ner-corpus
df1 = pd.read_csv("../Data/ner.csv")
df = df1.head(20)

df.head()

# Load model and tokenizer
model_name = "Babelscape/rebel-large"

tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSeq2SeqLM.from_pretrained(model_name)

def extract_relations(text):

    inputs = tokenizer(
        text, 
        return_tensors="pt",
        truncation=True,
        max_length=512
    )

    with torch.no_grad():
        outputs = model.generate(
            **inputs,
            max_length=256,
            num_beams=3
        )

    decoded = tokenizer.decode(outputs[0], skip_special_tokens=False)
    print (decoded) 
    
    return str(decoded)

def extract_triplet(text):
    """
    Extract (subject, relation, object) from REBEL-style output.
    """
    # Remove start/end tokens
    text = text.replace("<s>", "").replace("</s>", "").strip()

    # Regex pattern for REBEL format
    pattern = r"<triplet>([^<]+)<subj>([^<]+)<obj>([^<]+)"

    triplets = re.findall(pattern, text)
    #print (triplets)

    #for t in triplets:
    #    print(t)

    return triplets

def relation_extraction(strr):
    temp = extract_relations(strr)
    triples = extract_triplet(temp)
    return (triples)

strrs = "<s><triplet> Kalam <subj> Rameswaram <obj> place of birth</s>"
extract_triplet(strrs)

txt = 'Kalam was born in Rameswaram'
triplets = relation_extraction(txt)
triplets

txt = 'Rameswaram is where Kalam was born '
relation_extraction(txt)

# Apply relation extraction
df["triplets"] = df["Sentence"].apply(relation_extraction)

df.head()

df.to_csv('Relations.csv')
