==> WorkshopCodes/Shaastra2026-IITMadras/04-NLP/09-WebNews.ipynb <==
from GoogleNews import GoogleNews
from newspaper import Article
import pandas as pd

#pip install GoogleNews
#pip install newspaper3k
#pip install lxml_html_clean

from GoogleNews import GoogleNews

# Initialize GoogleNews object
googlenews = GoogleNews()

# Search for news by keyword
googlenews.search('technology')

# Get news results
results = googlenews.result()

for article in results:
    print(f"Title: {article['title']}")
    print(f"URL: {article['link']}")
    print()


def get_news_by_date(keyword, start_date, end_date):
    """
    start_date and end_date format: MM/DD/YYYY
    """
    googlenews = GoogleNews(lang='en', region='US')
    
    # Set date range
    googlenews.set_time_range(start_date, end_date)
    
    # Search keyword
    googlenews.search(keyword)
    
    # Get results
    results = googlenews.result()
    
    news_list = []
    for item in results:
        news_list.append({
            "title": item.get("title"),
            "date": item.get("date"),
            "source": item.get("media"),
            "description": item.get("desc"),
            "link": item.get("link")
        })
    
    return news_list

# Get silver news
keyword = "silver metal"
start_date = "20251201"
end_date = "20251231"

news = get_news_by_date(keyword, start_date, end_date)

for n in news:
    print(f"Title: {n['title']}")
    print(f"Date: {n['date']}")
    print(f"Source: {n['source']}")
    print(f"Link: {n['link']}")
    print("-" * 60)

def extract_text_from_url(url):
    article = Article(url)
    article.download()
    article.parse()
    
    return {
        "title": article.title,
        "authors": article.authors,
        "publish_date": article.publish_date,
        "text": article.text
    }


# Example usage
url = "https://finance.yahoo.com/news/prediction-ai-stock-could-first-152000439.html&ved=2ahUKEwjTmdPM1uCRAxX4WHADHYQxMpYQxfQBegQIBxAC&usg=AOvVaw1E5kYxXD-NfZtDk0rwdpDh"

data = extract_text_from_url(url)

print("Title:", data["title"])
print("Authors:", data["authors"])
print("Publish date:", data["publish_date"])
print("\nArticle text:\n")
print(data["text"][:1000])  # first 1000 characters


newstext = []

for n in news:
    url = n['link']
    url1, url2 = url.split('&ved=')
    print (url1)
    try:
        print(f"Link: {url1}")
        data = extract_text_from_url(url1)
        newstext.append(data["text"][:1000])   
    except: 
        print ('~~~~~~~~~')
        print ('cannot extract text from this link')
        print(f"Link: {url2}")
        print ('~~~~~~~~~')

print (newstext)

newsdf = pd.DataFrame(newstext)
newsdf.columns = ['NewsText']
newsdf.to_csv('NewsText.csv', index=False)
newsdf.head()
