==> WorkshopCodes/Shaastra2026-IITMadras/04-NLP/11-News-NER-LLMs.ipynb <==
import pandas as pd
import spacy
from spacy import displacy
import matplotlib.pyplot as plt
from pyvis.network import Network
import re
import spacy
from transformers import pipeline

# pip install ipywidgets

# Load spaCy's pre-trained model for NER
nlp = spacy.load("en_core_web_sm")

# Alternatively, you can use Hugging Face's transformers pipeline for NER
ner_pipeline = pipeline("ner", model="dbmdz/bert-large-cased-finetuned-conll03-english", tokenizer="dbmdz/bert-large-cased-finetuned-conll03-english")

# Load the small pre-trained NER model
#ner_pipeline = pipeline("ner", model="distilbert-base-cased", tokenizer="distilbert-base-cased")

# Load a SMALL model fine-tuned for NER
#ner_pipeline = pipeline("ner", model="distilbert-base-cased-finetuned-conll03-english, aggregation_strategy="simple"  # groups subword tokens)

# Load the small pre-trained NER model
#ner_pipeline = pipeline("ner", model="Minibase/NER-Small", tokenizer="Minibase/NER-Small")

# Sample text for NER
text = "Elon Musk is the CEO of SpaceX. He was born in South Africa in 1971."

# Run NER using the Hugging Face pipeline
ner_results = ner_pipeline(text)

# Print the recognized named entities
print("NER Results:")
for result in ner_results:
    print(f"{result['word']} - {result['entity']}")

# Sample text
text = "Barack Obama was born in Hawaii. He was elected president in 2008."

# Using spaCy for NER
print("spaCy NER Results:")
doc = nlp(text)
for ent in doc.ents:
    print(f"{ent.text} - {ent.label_}")

# Using Hugging Face NER pipeline
print("\nHugging Face NER Results:")
ner_results = ner_pipeline(text)
for result in ner_results:
    print(f"{result['word']} - {result['entity']}")

# 01 - Load Data

# data source: Kaggle - https://www.kaggle.com/datasets/naseralqaydeh/named-entity-recognition-ner-corpus
df = pd.read_csv("NewsText.csv")
df.head()

def clean_text(inptext):
    text = str(inptext)
    # replace new line
    text = re.sub(r'\n', ' ', text) 
    # replace special characters
    text = re.sub(r'[^a-zA-Z0-9 ]', '', text)
    return (text)

df['Article'] = df['NewsText'].apply(lambda x: clean_text(x))

df.head()

# Load spaCy's pre-trained English model
nlp = spacy.load("en_core_web_sm")

# Function to extract named entities from a text
def extract_entities(text):
    doc = nlp(text)
    entities = [(ent.text, ent.label_) for ent in doc.ents]
    return entities

# Apply NER on the text column
df['entities'] = df['Article'].apply(extract_entities)
df['LLM_entities'] = df['Article'].apply(ner_pipeline)

df.head()

# Save the results to a new CSV
df.to_csv("news_data_with_entities_2.csv", index=False)

print("NER extraction completed. Results saved to 'data_with_entities.csv'.")
